requirements.txt
livekit>=1.0.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
groq>=0.4.0
sounddevice>=0.4.6
numpy>=1.21.0
pydub>=0.25.1
requests>=2.26.0
livekit-agents>=0.1.0 